Added a main class that should run the whole program
