+++ src/main/java/com/intelliChef/view/RecipeUploadView.java
@@ -1,9 +1,7 @@
 package com.intelliChef.view;
 
 import com.intelliChef.adapters.UploadImageController;
-import com.intelliChef.data_access.GeminiAIClient;
 import com.intelliChef.entities.Ingredient;
-import com.intelliChef.use_case.analyzeImage.AnalyzeImageInteractor;
 
 import javax.swing.*;
 import java.awt.*;
@@ -60,20 +58,6 @@ public class RecipeUploadView extends JFrame {
 
         frame.getContentPane().add(uploadButton, BorderLayout.CENTER);
         frame.setVisible(true);
-    }
-
-    public static void main(String[] args) {
-        SwingUtilities.invokeLater(() -> {
-            GeminiAIClient vertexAIClient = new GeminiAIClient(
-                    "", // TODO: Change this to a valid API key
-                    "us-central1",
-                    "gemini-1.5-flash-001");
-            AnalyzeImageInteractor analyzeImageInteractor = new AnalyzeImageInteractor(vertexAIClient);
-            UploadImageController uploadImageController = new UploadImageController(analyzeImageInteractor);
-
-            RecipeUploadView recipeUploadView = new RecipeUploadView(uploadImageController);
-            recipeUploadView.setVisible(true);
-        });
     }
 }
